requirements/base.txt
django==1.8.7
django-extensions==1.7.4
django-filter==1.0.0
djangorestframework==3.5.3
django-compressor==2.1
csscompressor==0.9.4
mysqlclient==1.3.9
django-mysql==1.1.0
psycopg2==2.6.1
unipath==1.1



redis==2.10.5
django-celery==3.1.17
django-redis==4.5.0

Werkzeug==0.11.10